Rough prototype of client-side hot-reloading in Mojolicious
This needs server support because how would we send the notification
to the client. Also, this should become a module so that any
webserver can optionally hot-reload static (or generated) HTML.

Preferrably, you would maybe have all of your templates trigger the
hot reload too, at least in the "development" environment. I'm not so
sure about modifying the HTML in production :)